separate Flags Class
