+++ PYTHON/flags.py
@@ -1,0 +1,29 @@
+# *****************************************************************************
+# МОДУЛЬ:    flags
+# ФАЙЛ:      flags.py
+# ЗАГОЛОВОК: КЛАСС ФЛАГОВ (ДЛЯ КАЛЬКУЛЯТОРА)
+# ОПИСАНИЕ:  Описывает класс фдагов калькулятора
+# *****************************************************************************
+
+
+class Flags:
+	""" КЛАСС ДЛЯ ФЛАГОВ КАЛЬКУЛЯТОРА """
+
+	def __init__(self):
+		# clear display, 1-означает что новый ввод в A затрет старые результаты
+		# , а так же в комбинации с OP=1, что в А есть число
+		# , 0-ввод в регистр А разрешен
+		self.CD = True
+		# constant, 1-означает незавершенную операцию. Незавершенная операция, это последовательный
+		# ввод арифметических действий, пока не будет нажата EQ. В случае незавершенной операции
+		# вычисления ведуться как B op A. Если операция завершена, то предполагается что второго операнда
+		# нет и он берется из регистра В, и соответственно рассчитывается как A op B.
+		# Используется для выбора алгоритма АЛУ.
+		self.CONST = False
+		# equal - нажата ли клавиша "равно". Используется для выбора алгоритма АЛУ.
+		self.EQ = False
+
+	def clear(self):
+		self.CD = True
+		self.CONST = False
+		self.EQ = False
